refactor(globusflow): move to globus package
Move to separate Globus authentication package,
which can also include configuration for Globus.
